+++ ViewModels/Home/ProductDetailViewModel.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-
 namespace backend.ViewModels.Home
 {
     public class ProductDetailViewModel
@@ -8,8 +6,8 @@ namespace backend.ViewModels.Home
         public string Name { get; init; } = string.Empty;
         public string? CategoryName { get; init; }
         public string? Brand { get; init; }
-        public decimal Price { get; init; }
-        public decimal? OriginalPrice { get; init; }
+        public long Price { get; init; }
+        public long? OriginalPrice { get; init; }
         public double Discount { get; init; }
         public string ImageUrl { get; init; } = string.Empty;
         public string Description { get; init; } = string.Empty;
